separate freshness and coverage owners
