+++ crates/codestory-runtime/src/lib.rs
@@ -75,9 +75,7 @@ use codestory_workspace::{
 use crossbeam_channel::{Receiver, Sender, unbounded};
 use fs4::fs_std::FileExt;
 use parking_lot::Mutex;
-use serde::Deserialize;
 use std::cell::RefCell;
-use std::cmp::Ordering;
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::io::{self, BufRead};
 use std::ops::Deref;
@@ -89,8 +87,10 @@ use uuid::Uuid;
 
 mod affected;
 mod agent;
+mod index_freshness;
 mod publication;
 mod repo_text;
+mod route_coverage;
 mod search_intent;
 mod search_plan;
 mod search_publication;
@@ -100,8 +100,19 @@ mod search_terms;
 mod semantic_projection;
 mod snippets;
 mod workspace_state;
-use affected::*;
+use affected::{AffectedOperationIdentityIndex, IndexFreshnessObservation};
 pub use agent::{packet_step_trace_json, plan_packet};
+use index_freshness::{
+    CachedIndexFreshness, index_freshness_cache_ttl_secs, index_freshness_from_storage_with_policy,
+    index_freshness_observation_from_storage_with_identities, indexable_source_path_in_workspace,
+    indexable_source_path_with_root, open_existing_storage_for_read, open_storage_for_read,
+    storage_fingerprint, workspace_member_index_summaries, workspace_member_storage_summaries,
+};
+#[cfg(test)]
+use index_freshness::{
+    EXACT_SYMBOL_HYBRID_MAX_RESULTS_CAP, arm_after_index_freshness_fence_test_hook,
+    index_freshness_from_storage, indexable_source_path, not_checked_index_freshness,
+};
 #[cfg(test)]
 use publication::{
     PUBLICATION_TEST_FAULT, PublicationTestAction, PublicationTestBoundary,
@@ -112,6 +123,13 @@ use publication::{
     run_activation_search_before_revalidate_hook, run_full_refresh_staged_store_hook,
     run_incremental_staged_store_hook, run_semantic_projection_before_revalidate_hook,
     run_source_policy_after_plan_hook, run_source_policy_before_revalidate_hook,
+};
+#[cfg(test)]
+use route_coverage::compare_optional_confidence_desc;
+use route_coverage::{
+    RouteHandlerCandidate, compare_route_handler_candidates, framework_route_coverage_matrix,
+    language_support_summary_for_language, route_endpoint_adjusted_search_score,
+    route_endpoint_metadata_from_canonical, route_endpoint_metadata_from_openapi_label,
 };
 use search_intent::indexed_file_matches_language_filter;
 use search_publication::{
@@ -333,561 +351,6 @@ impl Drop for ActiveCoreReadGuard {
             active.replace(self.previous.take());
         });
     }
-}
-
-#[derive(Debug, Deserialize)]
-struct RouteEndpointCanonicalMetadata {
-    kind: String,
-    #[serde(default)]
-    framework: Option<String>,
-    method: String,
-    path: String,
-    #[serde(default)]
-    raw_path: Option<String>,
-    #[serde(default)]
-    params: Vec<String>,
-    #[serde(default)]
-    confidence: Option<String>,
-    #[serde(default)]
-    source_convention: Option<String>,
-    #[serde(default)]
-    provenance: Vec<String>,
-}
-
-fn route_endpoint_metadata_from_canonical(
-    raw: &str,
-    node: &GraphNode,
-    source_file: Option<&str>,
-) -> serde_json::Result<RouteEndpointMetadataDto> {
-    let canonical = serde_json::from_str::<RouteEndpointCanonicalMetadata>(raw)?;
-    let kind = match canonical.kind.as_str() {
-        "openapi_endpoint" => RouteEndpointKindDto::OpenapiEndpoint,
-        _ => RouteEndpointKindDto::FrameworkRoute,
-    };
-    Ok(RouteEndpointMetadataDto {
-        kind,
-        framework: canonical.framework,
-        method: canonical.method,
-        path: canonical.path,
-        raw_path: canonical.raw_path,
-        params: canonical.params,
-        source_file: source_file.map(ToOwned::to_owned),
-        line: node.start_line,
-        confidence: canonical.confidence,
-        source_convention: canonical.source_convention,
-        handler: None,
-        provenance: canonical.provenance,
-    })
-}
-
-fn route_endpoint_extraction_score_adjustment(canonical_id: Option<&str>) -> f32 {
-    let Some(raw) = canonical_id.and_then(|value| value.strip_prefix("route_endpoint:")) else {
-        return 0.0;
-    };
-    let Ok(canonical) = serde_json::from_str::<RouteEndpointCanonicalMetadata>(raw) else {
-        return 0.0;
-    };
-
-    if canonical.provenance.iter().any(|entry| {
-        matches!(
-            entry.as_str(),
-            "extraction:ast_indexed" | "extraction:tree_sitter_query"
-        )
-    }) {
-        return 0.025;
-    }
-    if canonical.provenance.iter().any(|entry| {
-        matches!(
-            entry.as_str(),
-            "extraction:text_only" | "extraction:lexical_fallback"
-        )
-    }) {
-        return -0.025;
-    }
-    0.0
-}
-
-fn route_endpoint_adjusted_search_score(score: f32, canonical_id: Option<&str>) -> f32 {
-    (score + route_endpoint_extraction_score_adjustment(canonical_id)).max(0.0)
-}
-
-fn route_endpoint_metadata_from_openapi_label(
-    label: &str,
-    node: &GraphNode,
-    source_file: Option<&str>,
-) -> Option<RouteEndpointMetadataDto> {
-    let (method, path) = label.split_once(' ')?;
-    Some(RouteEndpointMetadataDto {
-        kind: RouteEndpointKindDto::OpenapiEndpoint,
-        framework: None,
-        method: method.to_ascii_uppercase(),
-        path: path.to_string(),
-        raw_path: Some(path.to_string()),
-        params: route_endpoint_params(path),
-        source_file: source_file.map(ToOwned::to_owned),
-        line: node.start_line,
-        confidence: Some("schema".to_string()),
-        source_convention: Some("openapi".to_string()),
-        handler: None,
-        provenance: vec!["openapi".to_string()],
-    })
-}
-
-fn route_endpoint_params(path: &str) -> Vec<String> {
-    path.split('/')
-        .filter_map(|segment| {
-            let segment = segment.trim();
-            segment
-                .strip_prefix(':')
-                .or_else(|| {
-                    segment
-                        .strip_prefix('{')
-                        .and_then(|value| value.strip_suffix('}'))
-                })
-                .map(str::to_string)
-        })
-        .filter(|segment| !segment.is_empty())
-        .collect()
-}
-
-fn route_handler_certainty_rank(
-    certainty: Option<codestory_contracts::graph::ResolutionCertainty>,
-) -> u8 {
-    match certainty {
-        Some(codestory_contracts::graph::ResolutionCertainty::Certain) => 3,
-        Some(codestory_contracts::graph::ResolutionCertainty::Probable) => 2,
-        Some(codestory_contracts::graph::ResolutionCertainty::Uncertain) => 1,
-        None => 0,
-    }
-}
-
-#[derive(Debug, Clone)]
-struct RouteHandlerCandidate {
-    edge: GraphEdge,
-    target: GraphNode,
-}
-
-fn compare_optional_confidence_desc(left: Option<f32>, right: Option<f32>) -> Ordering {
-    let left = left.filter(|value| value.is_finite());
-    let right = right.filter(|value| value.is_finite());
-    match (left, right) {
-        (Some(left), Some(right)) => right.total_cmp(&left),
-        (Some(_), None) => Ordering::Less,
-        (None, Some(_)) => Ordering::Greater,
-        (None, None) => Ordering::Equal,
-    }
-}
-
-fn compare_route_handler_candidates(
-    left: &RouteHandlerCandidate,
-    right: &RouteHandlerCandidate,
-) -> Ordering {
-    // Persisted graph edges have unique IDs, so the edge ID tie-breaker makes
-    // this a total order for every valid route-handler candidate set.
-    compare_optional_confidence_desc(left.edge.confidence, right.edge.confidence)
-        .then_with(|| {
-            route_handler_certainty_rank(right.edge.certainty)
-                .cmp(&route_handler_certainty_rank(left.edge.certainty))
-        })
-        .then(left.target.canonical_id.cmp(&right.target.canonical_id))
-        .then(left.target.qualified_name.cmp(&right.target.qualified_name))
-        .then(
-            left.target
-                .serialized_name
-                .cmp(&right.target.serialized_name),
-        )
-        .then((left.target.kind as i32).cmp(&(right.target.kind as i32)))
-        .then(left.target.file_node_id.cmp(&right.target.file_node_id))
-        .then(left.target.start_line.cmp(&right.target.start_line))
-        .then(left.target.start_col.cmp(&right.target.start_col))
-        .then(left.target.end_line.cmp(&right.target.end_line))
-        .then(left.target.end_col.cmp(&right.target.end_col))
-        .then(left.target.id.cmp(&right.target.id))
-        .then(left.edge.id.cmp(&right.edge.id))
-        .then(left.edge.source.cmp(&right.edge.source))
-        .then(left.edge.target.cmp(&right.edge.target))
-        .then(left.edge.resolved_source.cmp(&right.edge.resolved_source))
-        .then(left.edge.resolved_target.cmp(&right.edge.resolved_target))
-        .then(left.edge.line.cmp(&right.edge.line))
-        .then(
-            left.edge
-                .callsite_identity
-                .cmp(&right.edge.callsite_identity),
-        )
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct FrameworkRouteCoverageEntry {
-    framework: &'static str,
-    language: &'static str,
-    status: &'static str,
-    coverage_evidence: &'static str,
-    confidence_floor: &'static str,
-    handler_link_support: &'static str,
-    unsupported_patterns: &'static [&'static str],
-    known_gaps: &'static [&'static str],
-    promotable: bool,
-}
-
-const FRAMEWORK_ROUTE_COVERAGE_ENTRIES: &[FrameworkRouteCoverageEntry] = &[
-    FrameworkRouteCoverageEntry {
-        framework: "express",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "tree_sitter_query_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_for_direct_handler_names_when_graph_resolution_succeeds",
-        unsupported_patterns: &[
-            "nested, injected, factory-returned, chained, and multi-target receivers are not promoted",
-            "dynamic paths and middleware arrays do not produce handler claims",
-            "inline and nested-call handlers do not produce name-based handler links",
-        ],
-        known_gaps: &[
-            "mounted app and router prefixes are not globally propagated",
-            "direct require('express')() construction is not promoted",
-        ],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "react-router",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &[
-            "loader/action route objects and nested config composition are partial",
-        ],
-        known_gaps: &["generated routes are not expanded"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "sveltekit",
-        language: "svelte/javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "file_convention",
-        handler_link_support: "probable_for_server_method_exports",
-        unsupported_patterns: &[
-            "route groups and advanced matcher params are normalized conservatively",
-        ],
-        known_gaps: &["layout load propagation is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "nextjs",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "file_convention",
-        handler_link_support: "probable_for_route_method_exports",
-        unsupported_patterns: &["middleware rewrites and route groups require source review"],
-        known_gaps: &["parallel routes and intercepting routes are not fully modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "remix",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "file_convention",
-        handler_link_support: "probable_for_loader_action_exports",
-        unsupported_patterns: &["route config composition and resource routes are partial"],
-        known_gaps: &["flat-route edge cases need more real-repo probes"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "astro",
-        language: "astro/javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "file_convention",
-        handler_link_support: "probable_for_endpoint_method_exports",
-        unsupported_patterns: &["redirects and integration-generated routes are not expanded"],
-        known_gaps: &["content collections do not create route nodes"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "nuxt",
-        language: "vue/javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "file_convention",
-        handler_link_support: "probable_for_server_handlers",
-        unsupported_patterns: &["route middleware and generated module routes are partial"],
-        known_gaps: &["custom router options are not evaluated"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "fastify",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "tree_sitter_query_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_for_direct_handler_names_when_graph_resolution_succeeds",
-        unsupported_patterns: &[
-            "nested, injected, factory-returned, chained, and multi-target receivers are not promoted",
-            "dynamic paths, method arrays, and nested route builders are not promoted",
-            "inline and wrapped handlers do not produce name-based handler links",
-        ],
-        known_gaps: &[
-            "register() prefix propagation is not modeled",
-            "schema and runtime middleware semantics are not evaluated",
-        ],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "koa",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_when_handler_name_resolves",
-        unsupported_patterns: &["router prefixes and middleware arrays are partial"],
-        known_gaps: &["mounted router prefixes are not globally propagated"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "hono",
-        language: "javascript/typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_when_handler_name_resolves",
-        unsupported_patterns: &["basePath/grouped routes are partial"],
-        known_gaps: &["OpenAPI helper generated routes are not expanded"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "nestjs",
-        language: "typescript",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "decorator",
-        handler_link_support: "probable_for_controller_method",
-        unsupported_patterns: &["global prefixes and dynamic decorator expressions are partial"],
-        known_gaps: &["module graph prefix propagation is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "django",
-        language: "python",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_when_handler_name_resolves",
-        unsupported_patterns: &["include() trees and namespaced URLConfs are not fully expanded"],
-        known_gaps: &["path converters beyond parameter names are not typed"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "flask",
-        language: "python",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "decorator",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["blueprint prefixes and dynamic method declarations are partial"],
-        known_gaps: &["method lists are not fully enumerated"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "fastapi",
-        language: "python",
-        status: "partial",
-        coverage_evidence: "validated_by_tree_sitter_query_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_for_decorated_handler",
-        unsupported_patterns: &[
-            "path= keyword arguments and escaped non-raw string literals are not exact routes",
-            "head/options/api_route/websocket decorators are not indexed",
-            "chained or multi-target FastAPI/APIRouter construction is not promoted",
-            "factory-returned or injected router receivers without module-scope construction are not claimed",
-        ],
-        known_gaps: &["include_router prefix propagation is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "rails",
-        language: "ruby",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["resource expansion is not fully enumerated"],
-        known_gaps: &["constraints/scopes are not expanded"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "laravel",
-        language: "php",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["controller arrays and route groups are partial"],
-        known_gaps: &["group middleware/prefix stacking is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "spring",
-        language: "java",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "annotation",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["class-level prefixes are not fully combined in every case"],
-        known_gaps: &["composed annotations are not expanded"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "aspnet",
-        language: "csharp",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "attribute",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["controller-level route templates are partial"],
-        known_gaps: &["minimal API grouping is not fully modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "axum",
-        language: "rust",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_when_handler_name_resolves",
-        unsupported_patterns: &["nested routers and stateful route composition are partial"],
-        known_gaps: &["Router::nest prefix propagation is limited"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "actix",
-        language: "rust",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "probable_when_handler_name_resolves",
-        unsupported_patterns: &["scoped services and macros are partial"],
-        known_gaps: &["web::scope prefix propagation is limited"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "rocket",
-        language: "rust",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "attribute",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["mount prefixes are not fully combined"],
-        known_gaps: &["rank/format route attributes are not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "gin",
-        language: "go",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed_text_only",
-        unsupported_patterns: &["router groups and middleware chains are partial"],
-        known_gaps: &["group middleware/prefix stacking is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "chi",
-        language: "go",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed_text_only",
-        unsupported_patterns: &["route groups and mounted subrouters are partial"],
-        known_gaps: &["group middleware/prefix stacking is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "echo",
-        language: "go",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed_text_only",
-        unsupported_patterns: &["group prefixes are partial"],
-        known_gaps: &["group middleware/prefix stacking is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "fiber",
-        language: "go",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed_text_only",
-        unsupported_patterns: &["group prefixes and mounted apps are partial"],
-        known_gaps: &["group middleware/prefix stacking is not modeled"],
-        promotable: true,
-    },
-    FrameworkRouteCoverageEntry {
-        framework: "vue-router",
-        language: "vue",
-        status: "partial",
-        coverage_evidence: "validated_by_indexer_regression",
-        confidence_floor: "heuristic",
-        handler_link_support: "not_claimed",
-        unsupported_patterns: &["imported route arrays and generated routes are partial"],
-        known_gaps: &["Nuxt file routes are reported separately as nuxt"],
-        promotable: true,
-    },
-];
-
-fn framework_route_coverage_matrix() -> Vec<FrameworkRouteCoverageDto> {
-    FRAMEWORK_ROUTE_COVERAGE_ENTRIES
-        .iter()
-        .map(framework_route_coverage_dto)
-        .collect()
-}
-
-fn framework_route_coverage_dto(entry: &FrameworkRouteCoverageEntry) -> FrameworkRouteCoverageDto {
-    FrameworkRouteCoverageDto {
-        framework: entry.framework.to_string(),
-        language: entry.language.to_string(),
-        status: entry.status.to_string(),
-        coverage_evidence: entry.coverage_evidence.to_string(),
-        confidence_floor: entry.confidence_floor.to_string(),
-        handler_link_support: entry.handler_link_support.to_string(),
-        unsupported_patterns: entry
-            .unsupported_patterns
-            .iter()
-            .map(|value| value.to_string())
-            .collect(),
-        known_gaps: entry
-            .known_gaps
-            .iter()
-            .map(|value| value.to_string())
-            .collect(),
-        promotable: entry.promotable,
-    }
-}
-
-struct LanguageSupportSummary {
-    support_mode: String,
-    evidence_tier: String,
-    claim_label: String,
-}
-
-fn language_support_summary_for_language(language: &str) -> LanguageSupportSummary {
-    language_support_profile_for_language_name(language)
-        .map(|profile| LanguageSupportSummary {
-            support_mode: profile.support_mode.as_str().to_string(),
-            evidence_tier: profile.evidence_tier.as_str().to_string(),
-            claim_label: profile.claim_label.to_string(),
-        })
-        .unwrap_or_else(|| LanguageSupportSummary {
-            support_mode: "unknown".to_string(),
-            evidence_tier: "unknown".to_string(),
-            claim_label: "no support claim recorded".to_string(),
-        })
 }
 
 #[derive(Clone)]
@@ -1706,573 +1169,6 @@ fn path_role_from_key(path: &str) -> IndexedFileRoleDto {
     }
 }
 
-const INDEX_FRESHNESS_INDEXED_FILE_CAP: usize = 25_000;
-const INDEX_FRESHNESS_CURRENT_FILE_CAP: usize = 25_000;
-const INDEX_FRESHNESS_SAMPLE_LIMIT: usize = 8;
-const INDEX_FRESHNESS_CACHE_DEFAULT_TTL_SECS: u64 = 60;
-#[cfg(test)]
-const EXACT_SYMBOL_HYBRID_MAX_RESULTS_CAP: usize = 80;
-
-#[cfg(test)]
-thread_local! {
-    static AFTER_INDEX_FRESHNESS_FENCE_TEST_HOOK: RefCell<Option<Box<dyn FnOnce()>>> =
-        const { RefCell::new(None) };
-}
-
-#[cfg(test)]
-fn arm_after_index_freshness_fence_test_hook(hook: impl FnOnce() + 'static) {
-    AFTER_INDEX_FRESHNESS_FENCE_TEST_HOOK.with(|slot| *slot.borrow_mut() = Some(Box::new(hook)));
-}
-
-#[cfg(test)]
-fn run_after_index_freshness_fence_test_hook() {
-    let hook = AFTER_INDEX_FRESHNESS_FENCE_TEST_HOOK.with(|slot| slot.borrow_mut().take());
-    if let Some(hook) = hook {
-        hook();
-    }
-}
-
-fn not_checked_index_freshness(
-    reason: impl Into<String>,
-    indexed_file_count: u32,
-    started_at: Instant,
-) -> IndexFreshnessDto {
-    IndexFreshnessDto {
-        status: IndexFreshnessStatusDto::NotChecked,
-        changed_file_count: 0,
-        new_file_count: 0,
-        removed_file_count: 0,
-        checked_file_count: 0,
-        indexed_file_count,
-        duration_ms: clamp_u128_to_u32(started_at.elapsed().as_millis()),
-        reason: Some(reason.into()),
-        samples: Vec::new(),
-    }
-}
-
-fn indexable_source_path(path: &Path) -> bool {
-    if path.to_str().is_some_and(|path| {
-        !Path::new(path).is_absolute()
-            && codestory_contracts::language_support::is_structural_source_path(path)
-            && codestory_contracts::language_support::structural_source_path_exclusion(path)
-                .is_some()
-    }) {
-        return false;
-    }
-    let tree_sitter_supported = path
-        .extension()
-        .and_then(|value| value.to_str())
-        .and_then(codestory_indexer::get_language_for_ext)
-        .is_some();
-    tree_sitter_supported
-        || codestory_indexer::template_pipeline::template_kind_for_path(path).is_some()
-        || codestory_indexer::structural::is_structural_candidate_path(path)
-        || codestory_indexer::is_text_only_candidate_path(path)
-        || looks_like_openapi_source_path(path)
-}
-
-fn indexable_source_path_in_workspace(root: &Path, path: &Path) -> bool {
-    let Some(relative) = codestory_workspace::workspace_relative_path(root, path) else {
-        return false;
-    };
-    indexable_source_path(&relative)
-}
-
-fn indexable_source_path_with_root(root: Option<&Path>, path: &Path) -> bool {
-    root.map_or_else(
-        || indexable_source_path(path),
-        |root| indexable_source_path_in_workspace(root, path),
-    )
-}
-
-fn looks_like_openapi_source_path(path: &Path) -> bool {
-    if !codestory_indexer::is_openapi_candidate_path(path) {
-        return false;
-    }
-    let Ok(source) = std::fs::read_to_string(path) else {
-        return true;
-    };
-    codestory_indexer::looks_like_openapi_schema(&source)
-}
-
-fn index_freshness_observation_from_storage(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    storage: &Storage,
-    policy: &SourceIndexPolicy,
-) -> IndexFreshnessObservation {
-    let mut identities = AffectedOperationIdentityIndex::native();
-    index_freshness_observation_from_storage_with_identities(
-        root,
-        workspace,
-        storage,
-        policy,
-        &mut identities,
-    )
-}
-
-fn index_freshness_observation_from_storage_with_identities<R>(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    storage: &Storage,
-    policy: &SourceIndexPolicy,
-    identities: &mut AffectedOperationIdentityIndex<R>,
-) -> IndexFreshnessObservation
-where
-    R: FnMut(&Path) -> io::Result<WorkspacePathIdentity>,
-{
-    let started_at = Instant::now();
-    match storage.has_incomplete_incremental_run() {
-        Ok(true) => {
-            return IndexFreshnessObservation::incomplete(IndexFreshnessDto {
-                status: IndexFreshnessStatusDto::Stale,
-                changed_file_count: 0,
-                new_file_count: 0,
-                removed_file_count: 0,
-                checked_file_count: 0,
-                indexed_file_count: 0,
-                duration_ms: clamp_u128_to_u32(started_at.elapsed().as_millis()),
-                reason: Some(
-                    "previous_incremental_run_incomplete_full_refresh_required".to_string(),
-                ),
-                samples: Vec::new(),
-            });
-        }
-        Ok(false) => {}
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to inspect incomplete index marker: {error}"),
-                0,
-                started_at,
-            ));
-        }
-    }
-    match storage.get_complete_index_publication() {
-        Ok(Some(publication)) => {
-            if let Err(error) = validate_structural_text_units(storage, &publication) {
-                return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                    format!(
-                        "structural text unit publication is incomplete: {}",
-                        error.message
-                    ),
-                    0,
-                    started_at,
-                ));
-            }
-            if let Err(error) =
-                validate_source_policy_exclusions(storage, root, &publication, policy)
-            {
-                return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                    format!(
-                        "source policy exclusion publication is incomplete: {}",
-                        error.message
-                    ),
-                    0,
-                    started_at,
-                ));
-            }
-        }
-        Ok(None) => {}
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to read complete core publication: {error}"),
-                0,
-                started_at,
-            ));
-        }
-    }
-    #[cfg(test)]
-    run_after_index_freshness_fence_test_hook();
-    let files = match storage.get_files() {
-        Ok(files) => files,
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to read indexed file inventory: {error}"),
-                0,
-                started_at,
-            ));
-        }
-    };
-    let indexed_file_count = clamp_usize_to_u32(files.len());
-    if files.is_empty() {
-        return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-            "no indexed file inventory is available yet",
-            indexed_file_count,
-            started_at,
-        ));
-    }
-    if files.len() > INDEX_FRESHNESS_INDEXED_FILE_CAP {
-        return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-            format!(
-                "indexed file inventory exceeds bounded freshness cap ({} > {})",
-                files.len(),
-                INDEX_FRESHNESS_INDEXED_FILE_CAP
-            ),
-            indexed_file_count,
-            started_at,
-        ));
-    }
-
-    let stored_files = match storage.files().inventory() {
-        Ok(files) => files,
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to read refresh inventory: {error}"),
-                indexed_file_count,
-                started_at,
-            ));
-        }
-    };
-    let removed_paths = files
-        .iter()
-        .map(|file| (file.id, file.path.clone()))
-        .collect::<HashMap<_, _>>();
-    let stored_policy_exclusions = match storage.get_source_policy_exclusions() {
-        Ok(exclusions) => exclusions,
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to read source policy exclusions: {error}"),
-                indexed_file_count,
-                started_at,
-            ));
-        }
-    };
-    let refresh_inputs = RefreshInputs {
-        stored_files,
-        policy_exclusions: stored_policy_exclusions
-            .iter()
-            .map(source_policy_exclusion_candidate)
-            .collect(),
-        inventory: Default::default(),
-    };
-    let refresh = match workspace.build_execution_outcome_bounded_with_policy(
-        &refresh_inputs,
-        INDEX_FRESHNESS_CURRENT_FILE_CAP,
-        policy,
-    ) {
-        Ok(refresh) => refresh,
-        Err(error) => {
-            return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-                format!("failed to check workspace inventory: {error}"),
-                indexed_file_count,
-                started_at,
-            ));
-        }
-    };
-    if refresh.refresh.inventory_outcome != WorkspaceInventoryOutcome::Complete {
-        let detail = refresh
-            .refresh
-            .inventory_issues
-            .first()
-            .map(|issue| format!("{}: {}", issue.path.display(), issue.message));
-        return IndexFreshnessObservation::incomplete(not_checked_index_freshness(
-            match detail {
-                Some(detail) => format!(
-                    "current workspace inventory is {:?}: {detail}",
-                    refresh.refresh.inventory_outcome
-                ),
-                None => format!(
-                    "current workspace inventory is {:?} (>{})",
-                    refresh.refresh.inventory_outcome, INDEX_FRESHNESS_CURRENT_FILE_CAP
-                ),
-            },
-            indexed_file_count,
-            started_at,
-        ));
-    }
-    let current_policy_exclusions = refresh.policy_exclusions;
-    let plan = refresh.refresh.plan;
-
-    let mut changed_file_count = 0u32;
-    let mut new_file_count = 0u32;
-    let mut samples = Vec::new();
-    for path in &plan.files_to_index {
-        let existing_indexed_file = plan.existing_file_ids.contains_key(path);
-        if !existing_indexed_file && !indexable_source_path_in_workspace(root, path) {
-            continue;
-        }
-        let kind = if existing_indexed_file {
-            changed_file_count = changed_file_count.saturating_add(1);
-            IndexFreshnessChangeKindDto::Changed
-        } else {
-            new_file_count = new_file_count.saturating_add(1);
-            IndexFreshnessChangeKindDto::New
-        };
-        if samples.len() < INDEX_FRESHNESS_SAMPLE_LIMIT {
-            samples.push(IndexFreshnessSampleDto {
-                kind,
-                path: runtime_relative_path(root, path),
-            });
-        }
-    }
-
-    let previous_policy_by_path = stored_policy_exclusions
-        .iter()
-        .map(|entry| (entry.normalized_path.as_str(), entry))
-        .collect::<HashMap<_, _>>();
-    let current_policy_paths = current_policy_exclusions
-        .iter()
-        .map(|entry| entry.normalized_path.as_str())
-        .collect::<HashSet<_>>();
-    let planned_paths = plan
-        .files_to_index
-        .iter()
-        .map(|path| runtime_relative_path(root, path))
-        .collect::<HashSet<_>>();
-    for exclusion in &current_policy_exclusions {
-        let kind = match previous_policy_by_path.get(exclusion.normalized_path.as_str()) {
-            Some(previous)
-                if previous.content_hash == exclusion.content_hash
-                    && previous.observed_size == exclusion.observed_size
-                    && previous.observed_unit_count == exclusion.observed_unit_count
-                    && previous.policy_version == exclusion.policy_version
-                    && previous.byte_cap == exclusion.byte_cap
-                    && previous.structural_unit_cap == exclusion.structural_unit_cap =>
-            {
-                continue;
-            }
-            Some(_) => {
-                changed_file_count = changed_file_count.saturating_add(1);
-                IndexFreshnessChangeKindDto::Changed
-            }
-            None => {
-                new_file_count = new_file_count.saturating_add(1);
-                IndexFreshnessChangeKindDto::New
-            }
-        };
-        if samples.len() < INDEX_FRESHNESS_SAMPLE_LIMIT {
-            samples.push(IndexFreshnessSampleDto {
-                kind,
-                path: exclusion.normalized_path.clone(),
-            });
-        }
-    }
-
-    let removed_policy_exclusions = stored_policy_exclusions
-        .iter()
-        .filter(|entry| {
-            !current_policy_paths.contains(entry.normalized_path.as_str())
-                && !planned_paths.contains(&entry.normalized_path)
-        })
-        .collect::<Vec<_>>();
-    let removed_file_count = clamp_usize_to_u32(
-        plan.files_to_remove
-            .len()
-            .saturating_add(removed_policy_exclusions.len()),
-    );
-    for removed_id in &plan.files_to_remove {
-        if samples.len() >= INDEX_FRESHNESS_SAMPLE_LIMIT {
-            break;
-        }
-        if let Some(path) = removed_paths.get(removed_id) {
-            samples.push(IndexFreshnessSampleDto {
-                kind: IndexFreshnessChangeKindDto::Removed,
-                path: runtime_relative_path(root, path),
-            });
-        }
-    }
-    for removed in removed_policy_exclusions {
-        if samples.len() >= INDEX_FRESHNESS_SAMPLE_LIMIT {
-            break;
-        }
-        samples.push(IndexFreshnessSampleDto {
-            kind: IndexFreshnessChangeKindDto::Removed,
-            path: removed.normalized_path.clone(),
-        });
-    }
-
-    let status = if changed_file_count == 0 && new_file_count == 0 && removed_file_count == 0 {
-        IndexFreshnessStatusDto::Fresh
-    } else {
-        IndexFreshnessStatusDto::Stale
-    };
-    let checked_file_count = indexed_file_count
-        .saturating_sub(removed_file_count)
-        .saturating_add(new_file_count);
-
-    for path in plan
-        .existing_file_ids
-        .keys()
-        .chain(plan.files_to_index.iter())
-    {
-        identities.record_admitted(path);
-    }
-    for path in &plan.files_to_index {
-        identities.record_stale(path);
-    }
-    for removed_id in &plan.files_to_remove {
-        let Some(path) = removed_paths.get(removed_id) else {
-            continue;
-        };
-        identities.record_stale(path);
-    }
-
-    let identity_gap_count = identities.freshness_identity_gap_count();
-    let identity_gap_sample = identities.freshness_identity_gap_sample();
-
-    IndexFreshnessObservation {
-        freshness: IndexFreshnessDto {
-            status,
-            changed_file_count,
-            new_file_count,
-            removed_file_count,
-            checked_file_count,
-            indexed_file_count,
-            duration_ms: clamp_u128_to_u32(started_at.elapsed().as_millis()),
-            reason: None,
-            samples,
-        },
-        inventory_complete: identity_gap_count == 0,
-        admitted_identities: identities.admitted_identities.clone(),
-        stale_identities: identities.stale_identities.clone(),
-        identity_gap_count,
-        identity_gap_sample,
-    }
-}
-
-fn index_freshness_from_storage_with_policy(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    storage: &Storage,
-    policy: &SourceIndexPolicy,
-) -> IndexFreshnessDto {
-    index_freshness_observation_from_storage(root, workspace, storage, policy).freshness
-}
-
-#[cfg(test)]
-fn index_freshness_from_storage(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    storage: &Storage,
-) -> IndexFreshnessDto {
-    index_freshness_from_storage_with_policy(
-        root,
-        workspace,
-        storage,
-        &SourceIndexPolicy::default(),
-    )
-}
-
-fn workspace_member_index_summaries(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    refresh_inputs: &RefreshInputs,
-    execution_plan: &RefreshExecutionPlan,
-) -> Vec<WorkspaceMemberIndexDto> {
-    workspace
-        .members()
-        .iter()
-        .map(|member| {
-            let absolute = if member.is_absolute() {
-                member.clone()
-            } else {
-                root.join(member)
-            };
-            let files_to_index = execution_plan
-                .files_to_index
-                .iter()
-                .filter(|path| path.starts_with(&absolute))
-                .count()
-                .min(u32::MAX as usize) as u32;
-            let indexed_files = refresh_inputs
-                .stored_files
-                .iter()
-                .filter(|file| file.path.starts_with(&absolute))
-                .count()
-                .min(u32::MAX as usize) as u32;
-            WorkspaceMemberIndexDto {
-                path: runtime_relative_path(root, &absolute),
-                files_to_index,
-                indexed_files,
-                file_count: None,
-                node_count: None,
-                edge_count: None,
-            }
-        })
-        .collect()
-}
-
-fn workspace_member_storage_summaries(
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    storage: &Storage,
-) -> Result<Vec<WorkspaceMemberIndexDto>, ApiError> {
-    if workspace.members().is_empty() {
-        return Ok(Vec::new());
-    }
-    let files = storage
-        .get_files()
-        .map_err(|e| ApiError::internal(format!("Failed to query member files: {e}")))?;
-    let nodes = storage
-        .get_nodes()
-        .map_err(|e| ApiError::internal(format!("Failed to query member nodes: {e}")))?;
-    let edges = storage
-        .get_edges()
-        .map_err(|e| ApiError::internal(format!("Failed to query member edges: {e}")))?;
-
-    let node_file_ids = nodes
-        .iter()
-        .map(|node| (node.id, node.file_node_id))
-        .collect::<HashMap<_, _>>();
-
-    Ok(workspace
-        .members()
-        .iter()
-        .map(|member| {
-            let absolute = if member.is_absolute() {
-                member.clone()
-            } else {
-                root.join(member)
-            };
-            let file_ids = files
-                .iter()
-                .filter(|file| file.path.starts_with(&absolute))
-                .map(|file| codestory_contracts::graph::NodeId(file.id))
-                .collect::<HashSet<_>>();
-            let file_count = file_ids.len().min(u32::MAX as usize) as u32;
-            let node_count = nodes
-                .iter()
-                .filter(|node| {
-                    file_ids.contains(&node.id)
-                        || node
-                            .file_node_id
-                            .is_some_and(|file_id| file_ids.contains(&file_id))
-                })
-                .count()
-                .min(u32::MAX as usize) as u32;
-            let edge_count = edges
-                .iter()
-                .filter(|edge| {
-                    edge.file_node_id
-                        .is_some_and(|file_id| file_ids.contains(&file_id))
-                        || node_file_ids
-                            .get(&edge.effective_source())
-                            .and_then(|file_id| *file_id)
-                            .is_some_and(|file_id| file_ids.contains(&file_id))
-                })
-                .count()
-                .min(u32::MAX as usize) as u32;
-            WorkspaceMemberIndexDto {
-                path: runtime_relative_path(root, &absolute),
-                files_to_index: 0,
-                indexed_files: file_count,
-                file_count: Some(file_count),
-                node_count: Some(node_count),
-                edge_count: Some(edge_count),
-            }
-        })
-        .collect())
-}
-
-#[derive(Debug, Clone)]
-struct CachedIndexFreshness {
-    root: PathBuf,
-    storage_path: PathBuf,
-    storage_fingerprint: String,
-    value: IndexFreshnessDto,
-    cached_at: Instant,
-}
-
 struct AppState {
     project_root: Option<PathBuf>,
     storage_path: Option<PathBuf>,
@@ -2284,73 +1180,6 @@ struct AppState {
     #[cfg(test)]
     #[allow(dead_code)]
     last_hybrid_instrumentation: Option<HybridSearchInstrumentation>,
-}
-
-fn index_freshness_cache_ttl_secs() -> u64 {
-    std::env::var("CODESTORY_INDEX_FRESHNESS_TTL_SECS")
-        .ok()
-        .and_then(|value| value.parse::<u64>().ok())
-        .filter(|ttl| *ttl > 0)
-        .unwrap_or(INDEX_FRESHNESS_CACHE_DEFAULT_TTL_SECS)
-}
-
-fn storage_fingerprint(path: &Path) -> String {
-    [
-        storage_path_fingerprint(path),
-        storage_path_fingerprint(&path.with_extension("db-wal")),
-        storage_path_fingerprint(&path.with_extension("db-shm")),
-    ]
-    .join("|")
-}
-
-fn storage_path_fingerprint(path: &Path) -> String {
-    let Ok(metadata) = std::fs::metadata(path) else {
-        return "missing".to_string();
-    };
-    let modified_ms = metadata
-        .modified()
-        .ok()
-        .and_then(|modified| modified.duration_since(UNIX_EPOCH).ok())
-        .map(|duration| duration.as_millis())
-        .unwrap_or_default();
-    format!("len:{}:mtime_ms:{modified_ms}", metadata.len())
-}
-
-fn open_storage_for_read(path: &Path) -> Result<Storage, ApiError> {
-    let requires_initialization = !path.exists()
-        || Storage::database_schema_version(path)
-            .map(|version| version != CURRENT_SCHEMA_VERSION)
-            .map_err(|error| {
-                ApiError::internal(format!("Failed to inspect storage schema: {error}"))
-            })?;
-    let storage = if requires_initialization {
-        Storage::open(path)
-    } else {
-        Storage::open_read_only(path)
-    };
-    storage.map_err(|error| ApiError::internal(format!("Failed to open storage: {error}")))
-}
-
-fn open_existing_storage_for_read(path: &Path) -> Result<Storage, ApiError> {
-    if !path.is_file() {
-        return Err(ApiError::new(
-            "project_unavailable",
-            "no complete project storage is available",
-        ));
-    }
-    let schema = Storage::database_schema_version(path).map_err(|error| {
-        ApiError::internal(format!("Failed to inspect storage schema: {error}"))
-    })?;
-    if schema != CURRENT_SCHEMA_VERSION {
-        return Err(ApiError::new(
-            "project_unavailable",
-            format!(
-                "project storage schema {schema} is not readable by runtime schema {CURRENT_SCHEMA_VERSION}"
-            ),
-        ));
-    }
-    Storage::open_read_only(path)
-        .map_err(|error| ApiError::internal(format!("Failed to open storage: {error}")))
 }
 
 fn publish_search_engine(
