+++ crates/codestory-runtime/src/path_resolution.rs
@@ -1,4 +1,6 @@
-use super::*;
+use super::{ApiError, AppController};
+use std::io;
+use std::path::{Path, PathBuf};
 
 pub(super) fn resolve_project_file_path(
     controller: &AppController,
