+++ crates/codestory-runtime/src/system_actions.rs
@@ -1,5 +1,7 @@
-use super::*;
+use super::{ApiError, SystemActionResponse, status_response};
 use std::ffi::OsString;
+use std::io;
+use std::path::Path;
 use std::process::Command;
 
 pub(super) fn expand_ide_template(
